+++ views/admin_portal.pug
@@ -46,7 +46,7 @@ block content
                     li
                       a(href='/luinet/datasets')= _("Thingpedia Datasets")
                     li
-                      a(href='/mturk')= _("MTurk Batches")
+                      a(href='/admin/mturk')= _("MTurk Batches")
 
       section.divider
         div.container
