+++ views/dev_sidebar.pug
@@ -19,5 +19,8 @@ nav#doc-sidebar
       if Config.WITH_THINGPEDIA === 'embedded'
         li(class=(section === 'alexa' ? 'current' : ''))
           a.doc-sidebar-subtitle(href='/developers/alexa')= _("Alexa Skills (Beta)")
+      if Config.WITH_LUINET === 'embedded'
+        li(class=(section === 'mturk' ? 'current' : ''))
+          a.doc-sidebar-subtitle(href='/developers/mturk')= _("MTurk Paraphrasing (Beta)")
